--- test/ansi.mjs
@@ -1,1 +1,0 @@
-export const ANSI = (x) => `\x1b[${x}m`;
