--- test/kleur.test.mjs
@@ -1,8 +1,0 @@
-import tap from 'tap';
-import c from '../index.mjs';
-
-tap.test('kleur', (t) => {
-	t.is(typeof c, 'object', 'exports an object');
-	t.ok(c.enabled, 'colors enabled by default');
-	t.end();
-});
